finished: adding tanstack router
I removed the App.tsx since we don't need it as the 'root' of the
project. I also configure the vite base path and matched it on the
main.tsx. I added index.tsx as a main page and remove unnecessary code
on __root.tsx. The routeTree.gen.ts will always be updated no matter
what.